changed when we read in our data
