+++ app/src/main/java/com/malmberg/matthew/projectgo/BioData.java
@@ -69,16 +69,4 @@ public class BioData {
         See = see;
     }
 
-    @Override
-    public String toString() {
-        return "BioData{" +
-                "name='" + name + '\'' +
-                ", price=" + price +
-                ", description='" + description + '\'' +
-                ", Eat=" + Eat +
-                ", Drink=" + Drink +
-                ", Listen=" + Listen +
-                ", See=" + See +
-                '}';
-    }
 }
