+++ app/src/main/res/layout/activity_options_list.xml
@@ -6,4 +6,10 @@
     android:layout_height="match_parent"
     tools:context="com.malmberg.matthew.projectgo.OptionsList">
 
+    <ListView
+        android:id="@+id/bio_list"
+        android:layout_width="368dp"
+        android:layout_height="495dp"
+        tools:layout_editor_absoluteX="8dp"
+        tools:layout_editor_absoluteY="8dp" />
 </android.support.constraint.ConstraintLayout>
